+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Opportunity sort and filter tables in query parameters
 - Remembering selected filters when switching tabs
+### Fixed
+- Direct deal for metaverse
 
 ## [1.18.0] - 2022-06-02
 ### Changed